quiet warnings

+++ UI/ObjectListWnd.cpp
@@ -1778,7 +1778,7 @@ private:
     void ButtonRightClicked(int column_id) {
         if (column_id < 0 || column_id >= static_cast<int>(m_controls.size()))
             return;
-        auto& clicked_button = m_controls[column_id+1];
+        auto& clicked_button = m_controls[static_cast<std::size_t>(column_id)+1u];
         if (!clicked_button)
             return;
